+++ source/helpers.js
@@ -203,3 +203,13 @@ export function zip(a, b)
 		return [a[index], b[index]]
 	})
 }
+
+export function get_language_from_locale(locale)
+{
+	const dash_index = locale.indexOf('-')
+	if (dash_index >= 0)
+	{
+		return locale.substring(0, dash_index)
+	}
+	return locale
+}